requirements.txt
langchain
transformers
pinecone-client
streamlit
openai
pdfplumber
beautifulsoup4
youtube-transcript-api
python-dotenv
torch
langchain_openai
langchain-community
langchain-huggingface
langchain_pinecone